remove reading of the internal temperature sensor
There's an internal temperature sensor but we do not know how to
convert the sensor readings to the actual temperature or the sensor
is simply /very/ inaccurate.

Up to now we never used these values so we can easily remove them
and increase readability of the program and reduce the LOC count.
If at a later point in time the temperature gets interesting again
we can add the code for it then.